Fix select star query for hive.
